requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
idna==3.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-whois==0.9.6
pytz==2025.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
